+++ Modules/Intent.Modules.Angular/Templates/Core/JsonPatches/AngularDotJson/OptionsPolyFillsPatch.cs
@@ -20,7 +20,7 @@ internal class OptionsPolyFillsPatch : IAngularJsonPatch
     }
     public AngularSettings.AngularVersionOptionsEnum AngularVersion { get; internal set; }
 
-    public bool Applicable() => AngularVersion == AngularSettings.AngularVersionOptionsEnum._19 || AngularVersion == AngularSettings.AngularVersionOptionsEnum._20;
+    public bool Applicable() => true;
 
     public void Apply(IDataFile file)
     {